+++ .evergreen/scripts/run_tests.py
@@ -214,3 +214,7 @@ def run() -> None:
 
 if __name__ == "__main__":
     run()
+
+
+# TODO - Make changes here to create a pytest marker that runs as desired
+#   Use encryption as example