+++ doc/changelog.rst
@@ -19,6 +19,7 @@ PyMongo 4.16 brings a number of changes including:
 - Use Zstandard support from the standard library for Python 3.14+, and use ``backports.zstd`` for older versions.
 - Fixed return type annotation for ``find_one_and_*`` methods on :class:`~pymongo.asynchronous.collection.AsyncCollection`
   and :class:`~pymongo.synchronous.collection.Collection` to include ``None``.
+- Added support for NumPy 1D-arrays in :class:`bson.binary.BinaryVector`.
 
 Changes in Version 4.15.5 (2025/XX/XX)
 --------------------------------------
@@ -42,6 +43,7 @@ See the `PyMongo 4.15.4 release notes in JIRA`_ for the list of resolved issues
 in this release.
 
 .. _PyMongo 4.15.4 release notes in JIRA: https://jira.mongodb.org/secure/ReleaseNote.jspa?projectId=10004&version=47237
+
 
 Changes in Version 4.15.3 (2025/10/07)
 --------------------------------------